refactor: Have one, not two files for querying termsets
